fixes revoke sponsorship cases for missing op type in enum
