+++ extension/src/popup/components/signTransaction/Operations/index.tsx
@@ -1,21 +1,9 @@
 import React from "react";
 import { Icon, IconButton } from "@stellar/design-system";
 import { useTranslation } from "react-i18next";
-import {
-  Asset,
-  buildInvocationTree,
-  Claimant,
-  LiquidityPoolAsset,
-  Operation,
-  scValToNative,
-  Signer,
-  SignerKeyOptions,
-  StrKey,
-  xdr,
-} from "stellar-sdk";
+import { Operation } from "stellar-sdk";
 
 import {
-  CLAIM_PREDICATES,
   FLAG_TYPES,
   OPERATION_TYPES,
   TRANSACTION_WARNING,
@@ -23,514 +11,18 @@ import {
 
 import { FlaggedKeys } from "types/transactions";
 
+import { truncateString, truncatedPoolId } from "helpers/stellar";
 import {
-  truncatedPoolId,
-  truncatedPublicKey,
-  truncateString,
-} from "helpers/stellar";
-import { getArgsForTokenInvocation } from "popup/helpers/soroban";
-
-import { KeyIdenticon } from "popup/components/identicons/KeyIdenticon";
-import { SorobanTokenInterface } from "@shared/constants/soroban/token";
+  KeyValueClaimants,
+  KeyValueInvokeHostFn,
+  KeyValueLine,
+  KeyValueList,
+  KeyValueSigner,
+  KeyValueSignerKeyOptions,
+  KeyValueWithPublicKey,
+  PathList,
+} from "./KeyVal";
 import "./styles.scss";
-
-const KeyValueList = ({
-  operationKey,
-  operationValue,
-}: {
-  operationKey: string;
-  operationValue: string | number | React.ReactNode;
-}) => (
-  <div className="Operations__pair" data-testid="OperationKeyVal">
-    <div>
-      {operationKey}
-      {operationKey ? ":" : null}
-    </div>
-    <div>{operationValue}</div>
-  </div>
-);
-
-const KeyValueWithPublicKey = ({
-  operationKey,
-  operationValue,
-}: {
-  operationKey: string;
-  operationValue: string;
-}) => (
-  <KeyValueList
-    operationKey={operationKey}
-    operationValue={<KeyIdenticon publicKey={operationValue} isSmall />}
-  />
-);
-
-const KeyValueRootInvocation = ({
-  invocation,
-}: {
-  invocation: xdr.SorobanAuthorizedInvocation;
-}) => {
-  const rootJson = buildInvocationTree(invocation);
-  return (
-    <div className="Operations__pair--smart-contract">
-      <div>Invocation:</div>
-      <div className="Operations__scValue">
-        <div>
-          <pre>
-            {JSON.stringify(
-              rootJson,
-              (_, val) => (typeof val === "bigint" ? val.toString() : val),
-              2,
-            )}
-          </pre>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const KeyValueSigner = ({ signer }: { signer: Signer }) => {
-  const { t } = useTranslation();
-
-  function renderSignerType() {
-    if ("ed25519PublicKey" in signer) {
-      return (
-        <KeyValueList
-          operationKey={t("Signer")}
-          operationValue={signer.ed25519PublicKey}
-        />
-      );
-    }
-    if ("sha256Hash" in signer) {
-      return (
-        <KeyValueList
-          operationKey={t("Signer")}
-          operationValue={formattedBuffer(signer.sha256Hash)}
-        />
-      );
-    }
-
-    if ("preAuthTx" in signer) {
-      return (
-        <KeyValueList
-          operationKey={t("Signer")}
-          operationValue={formattedBuffer(signer.preAuthTx)}
-        />
-      );
-    }
-
-    if ("ed25519SignedPayload" in signer) {
-      return (
-        <KeyValueList
-          operationKey={t("Signer")}
-          operationValue={signer.ed25519SignedPayload}
-        />
-      );
-    }
-    return <></>;
-  }
-
-  return (
-    <>
-      {renderSignerType()}
-      <KeyValueList
-        operationKey={t("Signer Weight")}
-        operationValue={signer.weight}
-      />
-    </>
-  );
-};
-
-const KeyValueLine = ({ line }: { line: Asset | LiquidityPoolAsset }) => {
-  const { t } = useTranslation();
-  if ("assetA" in line) {
-    return (
-      <>
-        <KeyValueList
-          operationKey={t("Asset A")}
-          operationValue={line.assetA}
-        />
-        <KeyValueList
-          operationKey={t("Asset B")}
-          operationValue={line.assetB}
-        />
-        <KeyValueList operationKey={t("Fee")} operationValue={line.fee} />
-      </>
-    );
-  }
-  return (
-    <KeyValueList operationKey={t("Asset Code")} operationValue={line.code} />
-  );
-};
-
-const KeyValueClaimants = ({ claimants }: { claimants: Claimant[] }) => {
-  const { t } = useTranslation();
-  return (
-    <div className="Operations--list--item" data-testid="OperationKeyVal">
-      <div>Claimants:</div>
-      {claimants.map((claimant) => (
-        <div key={claimant.destination + claimant.predicate.switch().name}>
-          <KeyValueWithPublicKey
-            operationKey={t("Destination")}
-            operationValue={claimant.destination}
-          />
-          <KeyValueList
-            operationKey={t("Predicate")}
-            operationValue={CLAIM_PREDICATES[claimant.predicate.switch().name]}
-          />
-        </div>
-      ))}
-    </div>
-  );
-};
-
-const KeyValueSignerKeyOptions = ({ signer }: { signer: SignerKeyOptions }) => {
-  const { t } = useTranslation();
-
-  if ("ed25519PublicKey" in signer) {
-    return (
-      <KeyValueList
-        operationKey={t("Signer Key")}
-        operationValue={signer.ed25519PublicKey}
-      />
-    );
-  }
-  if ("sha256Hash" in signer) {
-    return (
-      <KeyValueList
-        operationKey={t("Signer Sha256 Hash")}
-        operationValue={signer.sha256Hash}
-      />
-    );
-  }
-
-  if ("preAuthTx" in signer) {
-    return (
-      <KeyValueList
-        operationKey={t("Pre Auth Transaction")}
-        operationValue={signer.preAuthTx}
-      />
-    );
-  }
-
-  if ("ed25519SignedPayload" in signer) {
-    return (
-      <KeyValueList
-        operationKey={t("Signed Payload")}
-        operationValue={signer.ed25519SignedPayload}
-      />
-    );
-  }
-  return <></>;
-};
-
-const KeyValueInvokeHostFnArgs = ({ args }: { args: xdr.ScVal[] }) => (
-  <div className="Operations__pair" data-testid="OperationKeyVal">
-    <div>Parameters</div>
-    {args.map((arg) => (
-      <div>
-        <ScValByType scVal={arg} />
-      </div>
-    ))}
-  </div>
-);
-
-const ScValByType = ({ scVal }: { scVal: xdr.ScVal }) => {
-  switch (scVal.switch()) {
-    case xdr.ScValType.scvAddress(): {
-      const address = scVal.address();
-      const addressType = address.switch();
-      if (addressType.name === "scAddressTypeAccount") {
-        return StrKey.encodeEd25519PublicKey(address.accountId().ed25519());
-      }
-      return StrKey.encodeContract(address.contractId());
-    }
-
-    case xdr.ScValType.scvBool(): {
-      return scVal.b();
-    }
-
-    case xdr.ScValType.scvBytes(): {
-      return scVal.bytes().toString();
-    }
-
-    case xdr.ScValType.scvContractInstance(): {
-      const instance = scVal.instance();
-      return instance.executable().wasmHash().toString();
-    }
-
-    case xdr.ScValType.scvTimepoint():
-    case xdr.ScValType.scvDuration(): {
-      return scValToNative(scVal).toString();
-    }
-
-    case xdr.ScValType.scvError(): {
-      const error = scVal.error();
-      return `${error.contractCode()} - ${error.code().name}`;
-    }
-
-    case xdr.ScValType.scvI128():
-    case xdr.ScValType.scvI256():
-    case xdr.ScValType.scvI32():
-    case xdr.ScValType.scvI64():
-    case xdr.ScValType.scvU128():
-    case xdr.ScValType.scvU256():
-    case xdr.ScValType.scvU32():
-    case xdr.ScValType.scvU64(): {
-      return scValToNative(scVal).toString();
-    }
-
-    case xdr.ScValType.scvLedgerKeyNonce():
-    case xdr.ScValType.scvLedgerKeyContractInstance(): {
-      return scValToNative(scVal);
-    }
-
-    case xdr.ScValType.scvVec():
-    case xdr.ScValType.scvMap(): {
-      return JSON.stringify(scValToNative(scVal));
-    }
-
-    case xdr.ScValType.scvString():
-    case xdr.ScValType.scvSymbol(): {
-      const native = scValToNative(scVal);
-      if (native.constructor === "Uint8Array") {
-        return native.toString();
-      }
-      return native;
-    }
-
-    case xdr.ScValType.scvVoid(): {
-      return null;
-    }
-
-    default:
-      return null;
-  }
-};
-
-const KeyValueInvokeHostFn = ({ op }: { op: Operation.InvokeHostFunction }) => {
-  // TODO: render sub-invocation/auth
-  const { t } = useTranslation();
-  const authEntries = op.auth || [];
-  const hostfn = op.func;
-
-  function renderDetails() {
-    switch (hostfn.switch()) {
-      case xdr.HostFunctionType.hostFunctionTypeCreateContract(): {
-        const createContractArgs = hostfn.createContract();
-        const preimage = createContractArgs.contractIdPreimage();
-        const executable = createContractArgs.executable();
-        const executableType = executable.switch().name;
-
-        if (preimage.switch().name === "contractIdPreimageFromAddress") {
-          const preimageFromAddress = preimage.fromAddress();
-          const address = preimageFromAddress.address();
-          const salt = preimageFromAddress.salt().toString();
-
-          const addressType = address.switch();
-          if (addressType.name === "scAddressTypeAccount") {
-            const accountId = StrKey.encodeEd25519PublicKey(
-              address.accountId().ed25519(),
-            );
-            return (
-              <>
-                <KeyValueList
-                  operationKey={t("Invocation Type")}
-                  operationValue="Create Contract"
-                />
-                <KeyValueWithPublicKey
-                  operationKey={t("Account ID")}
-                  operationValue={accountId}
-                />
-                <KeyValueWithPublicKey
-                  operationKey={t("Salt")}
-                  operationValue={salt}
-                />
-                <KeyValueWithPublicKey
-                  operationKey={t("Executable Type")}
-                  operationValue={executableType}
-                />
-                <KeyValueWithPublicKey
-                  operationKey={t("Executable Wasm Hash")}
-                  operationValue={executable.wasmHash().toString()}
-                />
-              </>
-            );
-          }
-          const contractId = StrKey.encodeContract(address.contractId());
-          return (
-            <>
-              <KeyValueList
-                operationKey={t("Invocation Type")}
-                operationValue="Create Contract"
-              />
-              <KeyValueWithPublicKey
-                operationKey={t("Contract ID")}
-                operationValue={contractId}
-              />
-              <KeyValueWithPublicKey
-                operationKey={t("Salt")}
-                operationValue={salt}
-              />
-              <KeyValueWithPublicKey
-                operationKey={t("Executable Type")}
-                operationValue={executableType}
-              />
-              <KeyValueWithPublicKey
-                operationKey={t("Executable Wasm Hash")}
-                operationValue={executable.wasmHash().toString()}
-              />
-            </>
-          );
-        }
-
-        // contractIdPreimageFromAsset
-        const preimageFromAsset = preimage.fromAsset();
-        const preimageValue = preimageFromAsset.value()!;
-
-        return (
-          <>
-            <KeyValueList
-              operationKey={t("Invocation Type")}
-              operationValue="Create Contract"
-            />
-            {preimageFromAsset.switch().name === "assetTypeCreditAlphanum4" ||
-              (preimageFromAsset.switch().name ===
-                "assetTypeCreditAlphanum12" && (
-                <>
-                  <KeyValueList
-                    operationKey={t("Asset Code")}
-                    operationValue={(preimageValue as xdr.AlphaNum12)
-                      .assetCode()
-                      .toString()}
-                  />
-                  <KeyValueList
-                    operationKey={t("Issuer")}
-                    operationValue={StrKey.encodeEd25519PublicKey(
-                      (preimageValue as xdr.AlphaNum12).issuer().ed25519(),
-                    )}
-                  />
-                </>
-              ))}
-
-            <KeyValueWithPublicKey
-              operationKey={t("Executable Type")}
-              operationValue={executableType}
-            />
-            <KeyValueWithPublicKey
-              operationKey={t("Executable Wasm Hash")}
-              operationValue={executable.wasmHash().toString()}
-            />
-          </>
-        );
-      }
-
-      case xdr.HostFunctionType.hostFunctionTypeInvokeContract(): {
-        const invocation = hostfn.invokeContract();
-        const contractId = StrKey.encodeContract(
-          invocation.contractAddress().contractId(),
-        );
-        const fnName = invocation.functionName().toString();
-        const args = invocation.args();
-        const tokenMethods = [
-          SorobanTokenInterface.mint,
-          SorobanTokenInterface.transfer,
-        ];
-
-        if (tokenMethods.includes(fnName as SorobanTokenInterface)) {
-          const invokeParams = getArgsForTokenInvocation(fnName, args);
-          return (
-            <>
-              <KeyValueList
-                operationKey={t("Invocation Type")}
-                operationValue="Invoke Contract"
-              />
-              <KeyValueList
-                operationKey={t("Contract ID")}
-                operationValue={truncateString(contractId)}
-              />
-              <KeyValueList
-                operationKey={t("Function Name")}
-                operationValue={fnName}
-              />
-              <KeyValueList
-                operationKey={t("Amount")}
-                operationValue={invokeParams.amount.toString()}
-              />
-              <KeyValueWithPublicKey
-                operationKey={t("To")}
-                operationValue={invokeParams.to}
-              />
-              {fnName === SorobanTokenInterface.transfer && (
-                <KeyValueWithPublicKey
-                  operationKey={t("From")}
-                  operationValue={invokeParams.from}
-                />
-              )}
-              {authEntries.map((entry) => (
-                <KeyValueRootInvocation invocation={entry.rootInvocation()} />
-              ))}
-            </>
-          );
-        }
-        return (
-          <>
-            <KeyValueList
-              operationKey={t("Invocation Type")}
-              operationValue="Invoke Contract"
-            />
-            <KeyValueList
-              operationKey={t("Contract ID")}
-              operationValue={truncateString(contractId)}
-            />
-            <KeyValueList
-              operationKey={t("Function Name")}
-              operationValue={fnName}
-            />
-            <KeyValueInvokeHostFnArgs args={args} />
-          </>
-        );
-      }
-
-      case xdr.HostFunctionType.hostFunctionTypeUploadContractWasm(): {
-        const wasm = hostfn.wasm().toString();
-        return (
-          <>
-            <KeyValueList
-              operationKey={t("Invocation Type")}
-              operationValue="Upload Contract Wasm"
-            />
-            <KeyValueList operationKey={t("wasm")} operationValue={wasm} />
-          </>
-        );
-      }
-
-      default:
-        return <></>;
-    }
-  }
-  return renderDetails();
-};
-
-const PathList = ({ paths }: { paths: Asset[] }) => {
-  const { t } = useTranslation();
-
-  return (
-    <>
-      <div>{t("Paths")}: </div>
-      {paths.map(({ code, issuer }, i) => (
-        <div className="Operations--list--item" key={`${code} ${i + 1}`}>
-          <div>#{i + 1}</div>
-          <KeyValueList operationKey="Asset Code" operationValue={code} />
-          {issuer ? (
-            <KeyValueList
-              operationKey="Issuer"
-              operationValue={<KeyIdenticon publicKey={issuer} isSmall />}
-            />
-          ) : null}
-        </div>
-      ))}
-    </>
-  );
-};
 
 const UnsafeMaliciousWarning = ({
   isDestUnsafe,
@@ -609,9 +101,6 @@ const DestinationWarning = ({
   );
 };
 
-const formattedBuffer = (data: Buffer) =>
-  truncatedPublicKey(Buffer.from(data).toString("hex").toUpperCase());
-
 export const Operations = ({
   flaggedKeys,
   isMemoRequired,
@@ -646,7 +135,7 @@ export const Operations = ({
               flaggedKeys={flaggedKeys}
               isMemoRequired={isMemoRequired}
             />
-            <KeyValueWithPublicKey
+            <KeyValueList
               operationKey={t("Starting Balance")}
               operationValue={startingBalance}
             />
@@ -779,13 +268,23 @@ export const Operations = ({
       }
 
       case "manageSellOffer": {
-        const { offerId } = op;
+        const { offerId, selling, buying, price, amount } = op;
         return (
           <>
             <KeyValueList
               operationKey={t("Offer ID")}
               operationValue={offerId}
             />
+            <KeyValueList
+              operationKey={t("Selling")}
+              operationValue={selling.code}
+            />
+            <KeyValueList
+              operationKey={t("Buying")}
+              operationValue={buying.code}
+            />
+            <KeyValueList operationKey={t("Amount")} operationValue={amount} />
+            <KeyValueList operationKey={t("Price")} operationValue={price} />
           </>
         );
       }
@@ -829,49 +328,57 @@ export const Operations = ({
         } = op;
         return (
           <>
-            <KeyValueSigner signer={signer} />
-            <KeyValueList
-              operationKey={t("Inflation Destination")}
-              operationValue={inflationDest || ""}
-            />
-            <KeyValueList
-              operationKey={t("Home Domain")}
-              operationValue={homeDomain || ""}
-            />
-            <KeyValueList
-              operationKey={t("Inflation Destination")}
-              operationValue={inflationDest || ""}
-            />
-            <KeyValueList
-              operationKey={t("High Threshold")}
-              operationValue={highThreshold?.toString() || ""}
-            />
-            <KeyValueList
-              operationKey={t("Medium Threshold")}
-              operationValue={medThreshold?.toString() || ""}
-            />
-            <KeyValueList
-              operationKey={t("Low Threshold")}
-              operationValue={lowThreshold?.toString() || ""}
-            />
-            <KeyValueList
-              operationKey={t("Master Weight")}
-              operationValue={masterWeight?.toString() || ""}
-            />
-            <KeyValueList
-              operationKey={t("Set Flags")}
-              operationValue={
-                setFlags ? AuthorizationMapToDisplay[setFlags?.toString()] : ""
-              }
-            />
-            <KeyValueList
-              operationKey={t("Clear Flags")}
-              operationValue={
-                clearFlags
-                  ? AuthorizationMapToDisplay[clearFlags.toString()]
-                  : ""
-              }
-            />
+            {signer && <KeyValueSigner signer={signer} />}
+            {inflationDest && (
+              <KeyValueList
+                operationKey={t("Inflation Destination")}
+                operationValue={inflationDest}
+              />
+            )}
+            {homeDomain && (
+              <KeyValueList
+                operationKey={t("Home Domain")}
+                operationValue={homeDomain}
+              />
+            )}
+            {highThreshold && (
+              <KeyValueList
+                operationKey={t("High Threshold")}
+                operationValue={highThreshold?.toString()}
+              />
+            )}
+            {medThreshold && (
+              <KeyValueList
+                operationKey={t("Medium Threshold")}
+                operationValue={medThreshold?.toString()}
+              />
+            )}
+            {lowThreshold && (
+              <KeyValueList
+                operationKey={t("Low Threshold")}
+                operationValue={lowThreshold?.toString()}
+              />
+            )}
+            {masterWeight && (
+              <KeyValueList
+                operationKey={t("Master Weight")}
+                operationValue={masterWeight?.toString()}
+              />
+            )}
+            {setFlags && (
+              <KeyValueList
+                operationKey={t("Set Flags")}
+                operationValue={AuthorizationMapToDisplay[setFlags?.toString()]}
+              />
+            )}
+            {clearFlags && (
+              <KeyValueList
+                operationKey={t("Clear Flags")}
+                operationValue={
+                  AuthorizationMapToDisplay[clearFlags.toString()]
+                }
+              />
+            )}
           </>
         );
       }
@@ -880,7 +387,12 @@ export const Operations = ({
         const { source, type, limit, line } = op;
         return (
           <>
-            <KeyValueList operationKey={t("Source")} operationValue={source} />
+            {source && (
+              <KeyValueList
+                operationKey={t("Source")}
+                operationValue={source}
+              />
+            )}
             <KeyValueLine line={line} />
             <KeyValueList operationKey={t("Type")} operationValue={type} />
             <KeyValueList operationKey={t("Limit")} operationValue={limit} />
@@ -930,10 +442,12 @@ export const Operations = ({
         return (
           <>
             <KeyValueList operationKey={t("Name")} operationValue={name} />
-            <KeyValueList
-              operationKey={t("Value")}
-              operationValue={value?.toString()}
-            />
+            {value && (
+              <KeyValueList
+                operationKey={t("Value")}
+                operationValue={value?.toString()}
+              />
+            )}
           </>
         );
       }
@@ -964,7 +478,7 @@ export const Operations = ({
         return (
           <KeyValueList
             operationKey={t("Balance ID")}
-            operationValue={balanceId}
+            operationValue={truncateString(balanceId)}
           />
         );
       }
@@ -974,7 +488,7 @@ export const Operations = ({
         return (
           <KeyValueList
             operationKey={t("Sponsored ID")}
-            operationValue={sponsoredId}
+            operationValue={truncateString(sponsoredId)}
           />
         );
       }
@@ -983,105 +497,14 @@ export const Operations = ({
         const { source, type } = op;
         return (
           <>
-            <KeyValueList operationKey={t("Source")} operationValue={source} />
+            {source && (
+              <KeyValueList
+                operationKey={t("Source")}
+                operationValue={source}
+              />
+            )}
             <KeyValueList operationKey={t("Type")} operationValue={type} />
           </>
-        );
-      }
-
-      case "revokeSponsorship": {
-        // revoke trustline sponsorhip
-        if ("account" in op && "asset" in op) {
-          const { account, asset } = op;
-          return (
-            <>
-              <KeyValueList
-                operationKey={t("Account")}
-                operationValue={account}
-              />
-              {"liquidityPoolId" in asset && (
-                <KeyValueList
-                  operationKey={t("Liquidity Pool ID")}
-                  operationValue={truncatedPoolId(asset.liquidityPoolId)}
-                />
-              )}
-              {"code" in asset && (
-                <KeyValueList
-                  operationKey={t("Liquidity Pool ID")}
-                  operationValue={asset.code}
-                />
-              )}
-            </>
-          );
-        }
-        // revoke offer sponsorship
-        if ("seller" in op && "offerId" in op) {
-          const { seller, offerId } = op;
-          return (
-            <>
-              <KeyValueList
-                operationKey={t("Seller")}
-                operationValue={seller}
-              />
-              <KeyValueList
-                operationKey={t("Offer ID")}
-                operationValue={offerId}
-              />
-            </>
-          );
-        }
-        // revoke data sponsorship
-        if ("account" in op && "name" in op) {
-          const { account, name } = op;
-          return (
-            <>
-              <KeyValueList
-                operationKey={t("Account")}
-                operationValue={account}
-              />
-              <KeyValueList operationKey={t("Name")} operationValue={name} />
-            </>
-          );
-        }
-        //  revoke claimable sponsorship
-        if ("balanceId" in op) {
-          const { balanceId } = op;
-          return (
-            <KeyValueList
-              operationKey={t("Balance ID")}
-              operationValue={balanceId}
-            />
-          );
-        }
-        // revoke liquidity pool sponsorship
-        if ("liquidityPoolId" in op) {
-          const { liquidityPoolId } = op;
-          return (
-            <KeyValueList
-              operationKey={t("Liquidity Pool ID")}
-              operationValue={truncatedPoolId(liquidityPoolId)}
-            />
-          );
-        }
-        // revoke signer sponsorship
-        if ("signer" in op && "account" in op) {
-          const { account, signer } = op;
-          return (
-            <>
-              <KeyValueSignerKeyOptions signer={signer} />
-              <KeyValueList
-                operationKey={t("Account")}
-                operationValue={account}
-              />
-            </>
-          );
-        }
-
-        return (
-          <KeyValueList
-            operationKey={t("Account")}
-            operationValue={op.account}
-          />
         );
       }
 
@@ -1107,7 +530,7 @@ export const Operations = ({
         return (
           <KeyValueList
             operationKey={t("Balance ID")}
-            operationValue={balanceId}
+            operationValue={truncateString(balanceId)}
           />
         );
       }
@@ -1218,6 +641,95 @@ export const Operations = ({
       case "restoreFootprint":
       case "inflation":
       default: {
+        // OperationType is missing some types
+        // Issue:
+        const type = op.type as string;
+        if (type === "revokeTrustlineSponsorship") {
+          const _op = (op as unknown) as Operation.RevokeTrustlineSponsorship;
+          const { account, asset } = _op;
+          return (
+            <>
+              <KeyValueWithPublicKey
+                operationKey={t("Account")}
+                operationValue={account}
+              />
+              {"liquidityPoolId" in asset && (
+                <KeyValueList
+                  operationKey={t("Liquidity Pool ID")}
+                  operationValue={truncatedPoolId(asset.liquidityPoolId)}
+                />
+              )}
+              {"code" in asset && (
+                <KeyValueList
+                  operationKey={t("Liquidity Pool ID")}
+                  operationValue={asset.code}
+                />
+              )}
+            </>
+          );
+        }
+        if (type === "revokeAccountSponsorship") {
+          const _op = (op as unknown) as Operation.RevokeAccountSponsorship;
+          const { account } = _op;
+          return (
+            <KeyValueWithPublicKey
+              operationKey={t("Account")}
+              operationValue={account}
+            />
+          );
+        }
+        if (type === "revokeOfferSponsorship") {
+          const _op = (op as unknown) as Operation.RevokeOfferSponsorship;
+          const { seller, offerId } = _op;
+          return (
+            <>
+              <KeyValueWithPublicKey
+                operationKey={t("Seller")}
+                operationValue={seller}
+              />
+              <KeyValueList
+                operationKey={t("Offer ID")}
+                operationValue={offerId}
+              />
+            </>
+          );
+        }
+        if (type === "revokeDataSponsorship") {
+          const _op = (op as unknown) as Operation.RevokeDataSponsorship;
+          const { account, name } = _op;
+          return (
+            <>
+              <KeyValueWithPublicKey
+                operationKey={t("Account")}
+                operationValue={account}
+              />
+              <KeyValueList operationKey={t("Name")} operationValue={name} />
+            </>
+          );
+        }
+        if (type === "revokeClaimableBalanceSponsorship") {
+          const _op = (op as unknown) as Operation.RevokeClaimableBalanceSponsorship;
+          const { balanceId } = _op;
+          return (
+            <KeyValueList
+              operationKey={t("Balance ID")}
+              operationValue={truncateString(balanceId)}
+            />
+          );
+        }
+        if (type === "revokeSignerSponsorship") {
+          const _op = (op as unknown) as Operation.RevokeSignerSponsorship;
+          const { account, signer } = _op;
+          return (
+            <>
+              <KeyValueSignerKeyOptions signer={signer} />
+              <KeyValueWithPublicKey
+                operationKey={t("Account")}
+                operationValue={account}
+              />
+            </>
+          );
+        }
         return <></>;
       }
     }
